+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changed:
 * Made the UI a bit more clean and sleek and also make it mobile friendly.
 * Highlight when the king is in check.
 * Allow player to keep playing out the game after the puzzle is complete.
+* Added app configuration via environment variables - .env contains all of the available
+  configuration variables and can be used to set them.
 
 [1.3] - 2023-10-01
 Changed: